requirements.txt
faster-whisper>=1.0.3
sounddevice>=0.4.6
numpy>=1.24
keyboard>=0.13.5
pyperclip>=1.8.2
pyautogui>=0.9.54
pystray>=0.19.5
Pillow>=10.0.0
psutil>=5.9.0
openai>=1.0.0
